Criação e validação de Bearer token e endpoint de login. Adição da biblioteca AutoMapper e substituíção das converções de objeto para as funções da mesma. Modificação do swagger para autorização via Bearer  Token

+++ PontoX.Application/ApplicationServicePerfil.cs
@@ -11,47 +11,30 @@ namespace PontoX.Application
     public class ApplicationServicePerfil : IApplicationServicePerfil
     {
         private readonly IServicePerfil _service;
-        public ApplicationServicePerfil(IServicePerfil service)
+        private readonly IMapper _mapper;   
+        public ApplicationServicePerfil(IServicePerfil service, IMapper mapper)
         {
             _service = service;
+            _mapper = mapper;
         }
 
         public async Task<bool> AtualizarPerfil(PerfilRequest request)
         {
-            return await _service.Atualizar(new Perfil()
-            {
-                PerfilPaiId = request.PerfilPaiId,
-                Nome = request.Nome,
-                DataCriacao = request.DataCriacao,
-                DataModificacao = request.DataModificacao,
-                Id = request.Id
-            });
+            return await _service.Atualizar(_mapper.Map<Perfil>(request));
         }
 
         public async Task<bool> CadastrarPerfil(PerfilRequest request)
         {
             if (request.PerfilPaiId == null || request.PerfilPaiId == 0)
             {
-                return await _service.Adicionar(new Perfil()
-                {
-                    PerfilPaiId = request.PerfilPaiId,
-                    Nome = request.Nome,
-                    DataCriacao = request.DataCriacao,
-                    DataModificacao = request.DataModificacao
-                });
+                return await _service.Adicionar(_mapper.Map<Perfil>(request));
             }
             else
             {
                 var consulta = _service.ConsultarPorPerfilPaiId(request.PerfilPaiId);
                 if (consulta != null)
                 {
-                    return await _service.Adicionar(new Perfil()
-                    {
-                        PerfilPaiId = request.PerfilPaiId,
-                        Nome = request.Nome,
-                        DataCriacao = request.DataCriacao,
-                        DataModificacao = request.DataModificacao
-                    });
+                    return await _service.Adicionar(_mapper.Map<Perfil>(request));
                 }
                 else
                 {
@@ -64,18 +47,23 @@ namespace PontoX.Application
         public async Task<bool> ExcluirPerfil(long idCliente)
         {
             var perfilCliente = await _service.Consultar(idCliente);
-            var verification = await _service.ConsultarPerfilTemFilho(perfilCliente.Id);
-            if (verification.Count == 0)
-            {
-
-                _service.Remover(perfilCliente.Id);
-                return true;
-            }
-            else
+            if (perfilCliente == null)
             {
                 return false;
             }
-
+            else
+            {
+                var verification = await _service.ConsultarPerfilTemFilho(perfilCliente.Id);
+                if (verification.Count == 0)
+                {
+                    _service.Remover(perfilCliente.Id);
+                    return true;
+                }
+                else
+                {
+                    return false;
+                }
+            }
         }
 
         public async Task<List<PerfilResponse>> ListarPerfis()
@@ -87,20 +75,11 @@ namespace PontoX.Application
                 var perfilPai = new Perfil();
                 perfilPai = await _service.ConsultarPorPerfilPaiId(item.PerfilPaiId);
 
-                var perfil = new PerfilResponse()
-                {
-                    Nome = item.Nome,
-                    DataCriacao = item.DataCriacao,
-                    DataModificacao = item.DataModificacao,
-                };
+                var perfil = _mapper.Map<PerfilResponse>(perfilPai);
 
                 if (perfilPai != null)
                 {
-                    perfil.PerfilPai = new PerfilResponse()
-                    {
-                        Nome = perfilPai.Nome,
-                        DataCriacao = item.DataCriacao,
-                    };
+                    perfil.PerfilPai = _mapper.Map<PerfilResponse>(perfilPai);
                 }
                 perfilResponse.Add(perfil);
             }
